requirements.txt
gym==0.18.0
numpy==1.19.5
matplotlib==3.3.4
opencv-python==4.5.1.48
Pillow==7.2.0
stable-baselines[mpi]==2.10.1
tensorflow==1.15.0
alphavantage==0.0.11
requests==2.25.1
python-dateutil==2.8.1
